Better illustration of optimistic lock failures

+++ src/pg-lock-test/src/main/java/com/alexshabanov/service/UserService.java
@@ -5,7 +5,7 @@ import com.alexshabanov.service.domain.UserAccount;
 import java.math.BigDecimal;
 
 public interface UserService {
-    void setDelayMillis(long delayMillis);
+    void setDelayMillis(int delayMillis);
 
     void setUsePessimisticLock(boolean pessimisticLock);
